Clarify misunderstanding SESSIONID != Session ID
I find it more appropriate to use the term session identifiers at two places as it is a general term and could otherwise be read as "oh, ok this only applies to JSESSIONID and I am using JWT, so I am fine"

For consistency reasons I also changed it at another place which smells like session fixation which is not possible e.g. in the JWT world, normally.