Address feedback
* use TLS >= 1.2

Also, Bleichenbacher attack and friends (ROBOT)  due to PKCS#1 1.5 should not be anymore in the Description as one line. Reason: If you still have that somewhere enabled and attacks from 1998 / 2012 / 2018 work, your stack is that old that you have other severe problems. Also no developer probably knows what you're talking about when reading this line.

As for the following paragraph it's just a part of a line that can stay there but doesn't have to.